+++ src/main.ts
@@ -22,6 +22,8 @@ import { registerNoticeCommands } from "./commands/notice";
 import { registerQuestionnaireCommands } from "./commands/questionnaire";
 import { registerBoardroomCommands } from "./commands/boardroom";
 import { registerPersonalTodoCommands } from "./commands/personalTodo";
+import { registerPersonalTodoCommands } from "./commands/videoconference";
+import { registerVideoconferenceCommands } from "./commands/personalTodo";
 
 const program = new Command();
 
@@ -73,5 +75,6 @@ registerNoticeCommands(program);
 registerQuestionnaireCommands(program);
 registerBoardroomCommands(program);
 registerPersonalTodoCommands(program);
+registerVideoconferenceCommands(program);
 
 program.parse();